place/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
arrow==1.1.1
asgiref==3.4.1; python_version >= '3.6'
beautifulsoup4==4.9.3
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
diff-match-patch==20200713; python_version >= '2.7'
django-bootstrap-v5==1.0.4
django-import-export==2.5.0
django==3.2.6
et-xmlfile==1.1.0; python_version >= '3.6'
importlib-metadata==3.10.1; python_version < '3.9'
markuppy==1.14
odfpy==1.4.1
openpyxl==3.0.7
pymysql==1.0.2
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.1
pyyaml==5.4.1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.2.1; python_version >= '3.0'
sqlparse==0.4.1; python_version >= '3.5'
tablib[html,ods,xls,xlsx,yaml]==3.0.0; python_version >= '3.6'
xlrd==2.0.1
xlwt==1.3.0
zipp==3.5.0; python_version >= '3.6'
